Completed Posts and Comments card for Profile Page
