+++ client/src/components/ProfileLists/index.jsx
@@ -1,0 +1,4 @@
+import ProfilePosts from "./Posts";
+import ProfileComments from "./Comments";
+
+export { ProfilePosts, ProfileComments };
